+++ changelog.md
@@ -1,3 +1,9 @@
+## [Version 1.1.0] - 2024-10-20
+
+### Added
+- Object Pool System
+	- The Object Pool is specifically designed for Monobehaviors to be managed by the pools. Part of the advantage of the solution is that the pool can make objects aware of it so that there is fast lookup.
+
 ## [Version 1.0.1] - 2024-10-19
 
 ### Added